Move congestion into module
